breaking: move all the cli related code into a dedicated sub-pacakge
