+++ mxops/cli/__init__.py
@@ -1,0 +1,9 @@
+"""
+author: Etienne Wallet
+
+This sub-package contains the CLI of MxOps
+"""
+
+from mxops.cli.main_cli import cli_main
+
+__all__ = ["cli_main"]
